Registering User

+++ src/Utils/cloudinary.js
@@ -1,3 +1,5 @@
+import dotenv from "dotenv";
+dotenv.config();
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
@@ -6,6 +8,9 @@ cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
   api_secret: process.env.CLOUDINARY_API_SECRET 
+
+  
+  
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -21,9 +26,14 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+    console.error("Cloudinary Upload Error:", error);
+
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+}
 }
